strava_runner/requirements.txt
duckdb>=1.1.3
pandas>=2.2.3
pyarrow>=18.1.0
python-dotenv>=1.0.1
requests>=2.32.3
streamlit>=1.40.0
